Add sidekiq job to update statement for customer
This job runs the calculators and updates the customer statement.
The idea is that this job runs everytime an income/expenditure record is
created. At the moment there is not a requirement for income/expenditure
update actions but should that change in the future; this job can still
work within the classes that handle those actions